split colorscale methods into seperate files
